requirement.txt
python-dotenv
requests
IPython